Kurssit/Tietojenkäsittelytiede/avoin-ohpe-2012/viikko2-viikko2_08_tahtikehys/src/Tahtikehys.java
public class Tahtikehys {
    
    public static void kehysta(String teksti) {

        int pituus = teksti.length();
        String rivi = "****";

        for (int i = 0; i < pituus; i++) {
        rivi += "*";
        }
        
        System.out.println(rivi);
        System.out.println("* " + teksti + " *");
        System.out.println(rivi);
    
    }
    
    public static void main(String[] args) {
        kehysta("apina");
        kehysta("banaani");
        kehysta("cembalo");
    }
}
